third_party/blink/web_tests/platform/generic/http/tests/inspector-protocol/fenced-frame/handleJavaScriptDialog-fenced-frame.https-expected.txt

Tests that Page.handleJavaScriptDialog() in a fenced frame is not allowed.
Page.handleJavaScriptDialog() from a fenced frame:
PASS: This is only supported for top-level frames
